refactor: move setup to common handlers
